Switch to CMake - Taranis bootloader added. Tests done.
